Added perspective hitbox to enemy
This is just in case we want the enemy to be able to collide with the other objects in the room in a realistic way. As far as I can tell, the enemy can't leave the room area while following the player. May come in handy if we add knockback though to make sure the enemy isn't knocked out of the room.